set up running all go services with a single command
